Add support for merging script files in New-ConfigurationArtefact.ps1
This change adds support for merging script files with the New-ConfigurationArtefact function. The new parameters PreScriptMergePath and PostScriptMergePath can be used to specify the paths to the files that will be merged at the beginning and end of the script, respectively. If both the script blocks (PreScriptMerge and PostScriptMerge) and the file paths (PreScriptMergePath and PostScriptMergePath) are provided, the file paths will take precedence. The script files are read using the specified encoding. If the script content can be formatted using the Invoke-Formatter function, it will be formatted before merging. If formatting fails, the original script content will be used.